pyproject.toml
[tool.poetry]
name = "digit-classifier"
version = "0.1.0"
description = ""
authors = ["dhwanilditani <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21.0"
scikit-learn = "0.22.1"
pandas = "^1.3.0"
Django = "^3.2.5"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
